+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## 🛑 Breaking changes 🛑
 
 - Change semantic conventions for status (code, msg) as per specifications (#3872)
+- Add `pdata.NewTimestampFromTime`, deprecate `pdata.TimestampFromTime` (#3868)
 
 ## v0.33.0 Beta